# Final_Project/doc/API.yaml
openapi: 3.0.1
info:
  title: Management of members of private lessons
  version: 1.0.0
servers:
- url: https://private-lessons.com
...
paths:
  /adherent:
    get:
      tags:
      - Adherent
      summary: Get all adherents
      operationId: findAdh
      parameters:
      - name: page
        in: query
        required: true
        schema:
            $ref: '#/components/schemas/Page'
      - name: page_size
        in: query
        required: true
        schema:
            $ref: '#/components/schemas/PageSize'
      - name: first_name
        in: query
        description: Filter adherent by first name, case is ignored
        schema:
          type: string
      - name: last_name
        in: query
        description: Filter adherent by last name, case is ignored
        schema:
          type: string
      - name: status
        in: query
        description: Filter adherent by status, case is ignored
        schema:
          type: string
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Adherent'
        400:
          description: Bad request
          content: 
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BadRequest'

    put:
      tags:
      - Adherent
      summary: Create new adherent or update an existing Adherent
      operationId: CreateOrUpdateAdh
      description: Update adherent when `id_adherent` are provided, create them otherwise.
      requestBody:
        description: Adherent object that needs to be added in the request body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Adherent'
        required: true
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Adherent'
        400:
          description: Bad request
          content: 
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BadRequest'
  /adherent/{id_adherent}:
    get:
      tags:
      - Adherent
      summary: Get adherent by ID
      operationId: findAdhByID
      parameters:
      - name: id_adherent
        in: path
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Adherent'
        400:
          description: Bad request
          content: 
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BadRequest'

    delete:
      tags:
      - Adherent
      summary: Delete adherent by ID
      operationId: deleteAdh
      parameters:
      - name: id_adherent
        in: path
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
                example: Adherent succesfully deleted
        400:
          description: Bad request
          content: 
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BadRequest'
...
components:
  schemas:
    Adherent:
      type: object
      properties:
        id_adherent:
          type: integer
          format: int64
        first_name:
          type: string
          example: BEAR
        last_name:
          type: string
          example: Teddy
        phone:
          type: string
        email:
          type: string
          example: hey@gmail.com
        status:
          type: boolean
          enum:
            - true
            - false
...
    BadRequest:
      type: object
      properties:
        type:
          type: string
          example: BadRequestException
        first_name:
          type: string
          example: Bad Request
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`